+++ bin/name.js
@@ -58,10 +58,12 @@ function run(stack,options={}){
         set(++increment)
     }
 
+    config.stackNamePrefix = config.stackNamePrefix ? config.stackNamePrefix : 'QNA';
+
     if(options.prefix){
-        return `QNA-${full}`
+        return `${config.stackNamePrefix}-${full}`
     }else{
-        return `QNA-${full}-${increment}` 
+        return `${config.stackNamePrefix}-${full}-${increment}`
     }
 
     function set(value){