+++ lambda/fulfillment/lib/middleware/4_hook.js
@@ -1,7 +1,7 @@
 var _=require('lodash')
 var util=require('./util')
 
-module.exports=async function hook(req,res){
+module.exports = async function hook(req,res) {
     var arn=util.getLambdaArn(_.get(res.result,"l",""))
     
     if(arn){