+++ lambda/fulfillment/lib/middleware/5_assemble.js
@@ -20,6 +20,26 @@ function sms_hint(req,res) {
     return hint;
 }
 
+function resetAttributes(req,res) {
+    // Kendra attributes
+    let previous;
+    let prevQid;
+    let kendraResponsibleQid;
+    previous = _.get(req._event.sessionAttributes,"previous");
+    if (previous) {
+        let obj = JSON.parse(previous);
+        prevQid = obj.qid;
+    }
+    kendraResponsibleQid = _.get(res.session,"kendraResponsibleQid");
+    if ( (res.result === undefined || res.result.qid === undefined) || ( kendraResponsibleQid && (res.result.qid !== kendraResponsibleQid))) {
+        // remove any prior session attributes for kendra as they are no longer valid
+        if (res.session.kendraQueryId) delete res.session.kendraQueryId;
+        if (res.session.kendraIndexId) delete res.session.kendraIndexId;
+        if (res.session.kendraResultId) delete res.session.kendraResultId;
+        if (res.session.kendraResponsibleQid) delete res.session.kendraResponsibleQid;
+    }
+}
+
 module.exports=async function assemble(req,res){
     if(process.env.LAMBDA_LOG){
         await util.invokeLambda({
@@ -46,6 +66,9 @@ module.exports=async function assemble(req,res){
         _.get(res,'session',{}),
         x=>_.isString(x) ? x : JSON.stringify(x)
     )
+
+    resetAttributes(req,res);
+
     switch(req._type){
         case 'LEX':
             res.out=lex.assemble(req,res)
@@ -54,5 +77,6 @@ module.exports=async function assemble(req,res){
             res.out=alexa.assemble(req,res)
             break;
     }
+
     return {req,res}
 }