+++ lambda/lex-build/lib/bot.js
@@ -21,8 +21,15 @@ var _=require('lodash')
 var run=require('./run')
 
 module.exports=async function(versionobj,data){
-    data.intents[0].intentVersion=versionobj.intent_version
-    data.intents[1].intentVersion=versionobj.intentFallback_version
+    if (data.intents[0]['intentName'] && (data.intents[0]['intentName'].startsWith('fulfilment_'))) {
+        data.intents[0].intentVersion=versionobj.intent_version
+        if (data.intents.length > 1) {
+            data.intents[1].intentVersion=versionobj.intentFallback_version
+        }
+    } else {
+        data.intents[1].intentVersion=versionobj.intent_version
+        data.intents[0].intentVersion=versionobj.intentFallback_version
+    }
     delete data.status
     delete data.failureReason
     delete data.lastUpdatedDate