+++ templates/dev/bucket.js
@@ -35,7 +35,7 @@ module.exports=outputs('dev/bootstrap')
         "Handler": "index.handler",
         "MemorySize": "128",
         "Role": {"Fn::GetAtt": ["CFNLambdaRole","Arn"]},
-        "Runtime": "nodejs8.10",
+        "Runtime": "nodejs10.x",
         "Timeout": 60
       }
     },