+++ templates/examples/examples/py/Previous.py
@@ -3,11 +3,13 @@ import json
 import boto3
 import os
 
+stackoutputs = None
+stackname = os.getenv('CFSTACK')
+
 def handler(event, context):
 
-    #uncomment below if you want to see the JSON that is being passed to the Lambda Function
-    #jsondump = json.dumps(event)
-    #print(jsondump)
+    jsondump = json.dumps(event)
+    print(jsondump)
     
     # check we aren't calling this function before any document have been returned to the client and that 
     try:
@@ -37,20 +39,28 @@ def handler(event, context):
         response = json.loads(resp['Payload'].read())
         #uncomment below if you want to see the response 
         #print(json.dumps(response))
-        
-        # Do not call lambdafunction from the previous item if the link actually points to this previous function
-        if 'l' in response and response["l"].find(os.environ.get('AWS_LAMBDA_FUNCTION_NAME'))<=0:
+
+        # Do not call lambdafunction from the next item if the link points to ourselves
+        function_name = response.get('l', '')
+        if function_name != '' and function_name != 'QNA:ExamplePYTHONLambdaPrevious' and os.environ.get('AWS_LAMBDA_FUNCTION_NAME') not in function_name:
+            # This update will pull in standard qid content into the eventual result passed back up the stack
+            event = updateResult(event,response)
             if "args" in response:
                 event["res"]["result"]["args"] = response["args"]
             client = boto3.client('lambda')
+            targetname = response.get('l', '')
+            if targetname.startswith('arn') != True:
+                targetname = mapToArn(targetname, stackname)
             lhresp = client.invoke(
-                FunctionName = response["l"],
+                FunctionName = targetname,
                 Payload = json.dumps(event),
                 InvocationType = "RequestResponse"
             )
             # Because the payload is of a streamable type object, we must explicitly read it and load JSON
+            # Next merge in results of the LambdaHook execution
             event = updateLambdaHook(event,json.loads(lhresp['Payload'].read()),response)
         elif 'a' in response:
+            # No lambda hook to call so just merge in content from the target question
             event = updateResult(event,response)
                 # modify the event to make the previous question the redirected question that was just asked instead of "Next Question"
         else:
@@ -59,12 +69,23 @@ def handler(event, context):
 
         #uncomment line below if you want to see the final JSON before it is returned to the client
         # print(json.dumps(event))
-    # set the previous attribute back to the document qid that was previously returned since we don't want this document to be in history
-   # else:
-       # event["res"]["session"]["previous"] ={"qid":qidList,"a":navigationToJson["a"],"q":navigationToJson["q"],"next":navigationToJson["next"],"previous":[]}
 
     return event
 
+#maps a shortname to the full name via CF Output stack value
+def mapToArn(name,stack):
+    res = name
+    global stackoutputs
+    if stackoutputs is None:
+        cf = boto3.client('cloudformation')
+        r = cf.describe_stacks(StackName=stack)
+        stack, = r['Stacks']
+        stackoutputs = stack['Outputs']
+    for o in stackoutputs:
+        if name == 'QNA:' + o['OutputKey']:
+            res = o['OutputValue']
+            break
+    return res
 
 #update the event with the information if there is a Lambda hook
 def updateLambdaHook(event,hookEvent, response):
@@ -125,8 +146,6 @@ def updateResult(event, response):
     tempList.pop()
     event["res"]["session"]["previous"] ={"qid":response["qid"],"a":response["a"],"alt":response.get("alt",{}),"q":event["req"]["question"]}
     event["res"]["session"]["navigation"]={"next":response["next"],"previous":tempList,"hasParent":navigationToJson["hasParent"]}
-
-
 
     return event