+++ templates/master/elasticsearch/kibana/Dashboards.js
@@ -9,70 +9,20 @@ module.exports=[
       "title": "Default",
       "hits": 0,
       "description": "",
-      "panelsJSON":JSON.stringify([
-          {
-            "col": 9,
-            "id": "Client-types",
-            "panelIndex": 3,
-            "row": 1,
-            "size_x": 4,
-            "size_y": 4,
-            "type": "visualization"
-          },
-          {
-            "col": 1,
-            "id": "Requests",
-            "panelIndex": 4,
-            "row": 1,
-            "size_x": 8,
-            "size_y": 4,
-            "type": "visualization"
-          },
-          {
-            "col": 1,
-            "id": "Incorrect-feedback-wordcloud",
-            "panelIndex": 5,
-            "row": 9,
-            "size_x": 6,
-            "size_y": 3,
-            "type": "visualization"
-          },
-          {
-            "col": 7,
-            "id": "Correct-feedback-wordcloud",
-            "panelIndex": 6,
-            "row": 9,
-            "size_x": 6,
-            "size_y": 3,
-            "type": "visualization"
-          },
-          {
-            "col": 1,
-            "id": "Utterances",
-            "panelIndex": 7,
-            "row": 5,
-            "size_x": 6,
-            "size_y": 3,
-            "type": "visualization"
-          },
-          {
-            "col": 7,
-            "id": "No-Hits",
-            "panelIndex": 8,
-            "row": 5,
-            "size_x": 6,
-            "size_y": 3,
-            "type": "visualization"
-          }
-        ]),
+      "panelsJSON": "[{\"col\":9,\"id\":\"Client-types\",\"panelIndex\":3,\"row\":1,\"size_x\":4,\"size_y\":4,\"type\":\"visualization\"},{\"col\":1,\"id\":\"Requests\",\"panelIndex\":4,\"row\":1,\"size_x\":8,\"size_y\":4,\"type\":\"visualization\"},{\"col\":1,\"id\":\"Incorrect-feedback-wordcloud\",\"panelIndex\":5,\"row\":24,\"size_x\":12,\"size_y\":6,\"type\":\"visualization\"},{\"col\":1,\"id\":\"Correct-feedback-wordcloud\",\"panelIndex\":6,\"row\":18,\"size_x\":12,\"size_y\":6,\"type\":\"visualization\"},{\"col\":1,\"id\":\"Utterances\",\"panelIndex\":7,\"row\":5,\"size_x\":12,\"size_y\":6,\"type\":\"visualization\"},{\"col\":1,\"id\":\"No-Hits\",\"panelIndex\":8,\"row\":11,\"size_x\":12,\"size_y\":7,\"type\":\"visualization\"}]",
       "optionsJSON": "{\"darkTheme\":false}",
       "uiStateJSON": "{}",
-      "version": 1,
-      "timeRestore": false,
-      "timeFrom":"now/w",
-      "timeTo":"now/w",
+      "version": "1",
+      "timeRestore": "false",
+      "timeTo": "now/w",
+      "timeFrom": "now/w",
+      "refreshInterval": {
+        "display": "Off",
+        "pause": false,
+        "value": 0
+      },
       "kibanaSavedObjectMeta": {
-        "searchSourceJSON": "{\"filter\":[{\"query\":{\"query_string\":{\"query\":\"*\",\"analyze_wildcard\":true}}}]}"
+        "searchSourceJSON": "{\"filter\":[{\"query\":{\"query_string\":{\"analyze_wildcard\":true,\"query\":\"*\"}}}]}"
       }
     }
   },
@@ -82,7 +32,7 @@ module.exports=[
     "_index":".kibana",
     "_source": {
       "title": "Logged Utterances",
-      "visState": "{\"title\":\"Logged Utterances\",\"type\":\"tagcloud\",\"params\":{\"scale\":\"linear\",\"orientation\":\"single\",\"minFontSize\":16,\"maxFontSize\":100,\"hideLabel\":true},\"aggs\":[{\"id\":\"1\",\"enabled\":true,\"type\":\"count\",\"schema\":\"metric\",\"params\":{}},{\"id\":\"2\",\"enabled\":true,\"type\":\"terms\",\"schema\":\"segment\",\"params\":{\"field\":\"utterance.keyword\",\"size\":1000,\"order\":\"desc\",\"orderBy\":\"1\"}}],\"listeners\":{}}",
+      "visState": "{\"title\":\"Logged Utterances\",\"type\":\"tagcloud\",\"params\":{\"scale\":\"linear\",\"orientation\":\"single\",\"minFontSize\":16,\"maxFontSize\":50,\"hideLabel\":true},\"aggs\":[{\"id\":\"1\",\"enabled\":true,\"type\":\"count\",\"schema\":\"metric\",\"params\":{}},{\"id\":\"2\",\"enabled\":true,\"type\":\"terms\",\"schema\":\"segment\",\"params\":{\"field\":\"utterance.keyword\",\"size\":1000,\"order\":\"desc\",\"orderBy\":\"1\"}}],\"listeners\":{}}",
       "uiStateJSON": "{}",
       "description": "",
       "version": 1,
@@ -97,12 +47,12 @@ module.exports=[
     "_index":".kibana",
     "_source": {
       "title": "No Hits",
-      "visState": "{\"title\":\"No Hits\",\"type\":\"tagcloud\",\"params\":{\"scale\":\"linear\",\"orientation\":\"single\",\"minFontSize\":16,\"maxFontSize\":100,\"hideLabel\":true},\"aggs\":[{\"id\":\"1\",\"enabled\":true,\"type\":\"count\",\"schema\":\"metric\",\"params\":{}},{\"id\":\"2\",\"enabled\":true,\"type\":\"terms\",\"schema\":\"segment\",\"params\":{\"field\":\"utterance.keyword\",\"size\":1000,\"order\":\"desc\",\"orderBy\":\"1\"}}],\"listeners\":{}}",
+      "visState": "{\"title\":\"No Hits\",\"type\":\"tagcloud\",\"params\":{\"scale\":\"linear\",\"orientation\":\"single\",\"minFontSize\":16,\"maxFontSize\":50,\"hideLabel\":true},\"aggs\":[{\"id\":\"1\",\"enabled\":true,\"type\":\"count\",\"schema\":\"metric\",\"params\":{}},{\"id\":\"2\",\"enabled\":true,\"type\":\"terms\",\"schema\":\"segment\",\"params\":{\"field\":\"utterance.keyword\",\"size\":1000,\"order\":\"desc\",\"orderBy\":\"1\"}}],\"listeners\":{}}",
       "uiStateJSON": "{}",
       "description": "",
       "version": 1,
       "kibanaSavedObjectMeta": {
-        "searchSourceJSON": {"Fn::Sub":"{\"index\":\"${ESVar.MetricsIndex}\",\"query\":{\"query_string\":{\"query\":\"entireResponse.got_hits=0\",\"analyze_wildcard\":true}},\"filter\":[]}"}
+        "searchSourceJSON": {"Fn::Sub":"{\"index\":\"${ESVar.MetricsIndex}\",\"query\":{\"query_string\":{\"query\":\"entireResponse.got_hits:0\",\"analyze_wildcard\":true}},\"filter\":[]}"}
       }
     }
   },
@@ -253,7 +203,7 @@ module.exports=[
           ],
           "listeners": {},
           "params": {
-            "maxFontSize": 72,
+            "maxFontSize": 50,
             "minFontSize": 18,
             "orientation": "single",
             "scale": "linear"
@@ -299,7 +249,7 @@ module.exports=[
           ],
           "listeners": {},
           "params": {
-            "maxFontSize": 72,
+            "maxFontSize": 50,
             "minFontSize": 18,
             "orientation": "single",
             "scale": "linear"